Ransomware/Ransomware_code/requirements.txt
cryptography==41.0.5
customtkinter==5.2.1
Pillow==10.1.0
Pillow==10.4.0
